Rename fonts for epub renderer
This fixes rendering issues on the Kobo epub renderer which is used
when Kobo is not reading a kepub file, but a regular epub file.

It seems that the fallback fonts are loaded differently when loading
a regular .epub file, and fonts ending in -Roman.ttf are not loaded
correctly. This is resolved when the usual naming policy is applied.

Only NV Charter and NV Palatium used the incorrect file names.
Only users reading epub files on Kobo devices would have run into
this issue, but this is a silly one.

Also added addendum about kepubify to the README related to the
different renderers.